# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Małopolska Social Innovation Hub — Matchmaking API",
    "version": "1.0.0",
    "description": "AI-assisted matchmaking between a described social problem and existing social innovations, documented cases and contextual evidence. Relevance means suitability for the described need, not a probability of success."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current host"
    }
  ],
...
  "paths": {
...
    "/api/v1/tests": {
      "get": {
        "tags": [
          "Tester"
        ],
        "operationId": "listInnovationTests",
        "summary": "List innovation tests open to testers",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/TestStatus"
            }
          },
          {
            "name": "innovationId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 200
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 20
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 10000,
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of innovation tests.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Tester"
        ],
        "operationId": "createInnovationTest",
        "summary": "Create an innovation test (operator)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTestRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Test created.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "message",
                    "data"
                  ],
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/InnovationTest"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tests/{testId}": {
      "get": {
        "tags": [
          "Tester"
        ],
        "operationId": "getInnovationTest",
        "summary": "Get test details, application count and remaining slots",
        "parameters": [
          {
            "name": "testId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Test details.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/InnovationTest"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Test not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TestStatus": {
        "type": "string",
        "enum": [
          "recruiting",
          "active",
          "completed",
          "cancelled"
        ]
      },
...
      "InnovationTest": {
        "type": "object",
        "required": [
          "id",
          "innovationId",
          "title",
          "status",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "innovationId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "instructions": {
            "type": [
              "string",
              "null"
            ]
          },
          "location": {
            "type": [
              "string",
              "null"
            ]
          },
          "maxTesters": {
            "type": [
              "integer",
              "null"
            ]
          },
          "startAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "endAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "status": {
            "$ref": "#/components/schemas/TestStatus"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "innovationTitle": {
            "type": [
              "string",
              "null"
            ]
          },
          "innovationSummary": {
            "type": [
              "string",
              "null"
            ]
          },
          "applicationsCount": {
            "type": "integer",
            "minimum": 0
          },
          "acceptedCount": {
            "type": "integer",
            "minimum": 0
          },
          "slotsLeft": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 0
          }
        }
      },
      "TestListResponse": {
        "type": "object",
        "required": [
          "total",
          "count",
          "limit",
          "offset",
          "data"
        ],
        "properties": {
          "total": {
            "type": "integer",
            "minimum": 0
          },
          "count": {
            "type": "integer",
            "minimum": 0
          },
          "limit": {
            "type": "integer",
            "minimum": 1
          },
          "offset": {
            "type": "integer",
            "minimum": 0
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InnovationTest"
            }
          }
        }
      },
      "CreateTestRequest": {
        "type": "object",
        "required": [
          "innovationId",
          "title"
        ],
        "additionalProperties": false,
        "properties": {
          "innovationId": {
            "type": "string",
            "maxLength": 200
          },
          "title": {
            "type": "string",
            "maxLength": 300
          },
          "description": {
            "type": "string",
            "maxLength": 5000
          },
          "instructions": {
            "type": "string",
            "maxLength": 5000
          },
          "location": {
            "type": "string",
            "maxLength": 300
          },
          "maxTesters": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100000
          },
          "startAt": {
            "type": "string",
            "format": "date-time"
          },
          "endAt": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "$ref": "#/components/schemas/TestStatus"
          }
        }
      },
...
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message",
              "requestId",
              "retryable"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string",
                "description": "Safe Polish message."
              },
              "requestId": {
                "type": "string"
              },
              "retryable": {
                "type": "boolean"
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}